refactor: unify tooltip pattern, extract types, add chat barrel
- Migrate all SimpleTooltip usages to TooltipActionButton (trigger+icon props);
  delete simple-tooltip.tsx — single tooltip primitive in codebase
- Extract UsedContextChunk named type to chat.ts; remove duplicate
  definitions of UsedContextChunk and RagSource from build-rag-context.ts,
  re-export them from @/types instead
- Add src/features/chat/components/index.ts barrel for external consumers;
  update 5 files to import from barrel instead of deep paths
- Keep Chat import in sidepanel as direct path to avoid transitive
  Chrome API imports in test environment
