+++ src/features/chat/components/char-count.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next"
 
-import { SimpleTooltip } from "@/components/ui/simple-tooltip"
+import { TooltipActionButton } from "@/components/actions"
 
 export const CharCount = ({ count }: { count: number }) => {
   const { t } = useTranslation()
@@ -8,17 +8,18 @@ export const CharCount = ({ count }: { count: number }) => {
   if (count === 0) return null
 
   return (
-    <SimpleTooltip
-      content={
+    <TooltipActionButton
+      trigger={
+        <div className="cursor-default font-mono text-xs text-muted-foreground" />
+      }
+      tooltip={
         <p>
           {count} {t("chat.input.chars")}
         </p>
       }
-      side="top"
-      triggerRender={
-        <div className="cursor-default font-mono text-xs text-muted-foreground" />
-      }>
-      {count}
-    </SimpleTooltip>
+      tooltipSide="top"
+      showLabel
+      label={count}
+    />
   )
 }
