+++ src/features/chat/components/chat-message-footer.tsx
@@ -4,7 +4,6 @@ import {
   type ActionMenuItemConfig,
   TooltipActionButton
 } from "@/components/actions"
-import { SimpleTooltip } from "@/components/ui/simple-tooltip"
 import { chatIconBtnCls } from "@/features/chat/lib/chat-styles"
 import { ChatSessionActions } from "@/features/sessions/components/chat-session-actions"
 import {
@@ -189,12 +188,16 @@ export const ChatMessageFooter = ({
         )}
 
         {!isUser && msg.model && !isLoading && (
-          <SimpleTooltip content={t("chat.actions.switch_model_tooltip")}>
-            <RegenerateButton
-              model={msg.model}
-              onSelectModel={(model) => onRegenerate?.(model)}
-            />
-          </SimpleTooltip>
+          <TooltipActionButton
+            trigger={<span />}
+            tooltip={t("chat.actions.switch_model_tooltip")}
+            icon={
+              <RegenerateButton
+                model={msg.model}
+                onSelectModel={(model) => onRegenerate?.(model)}
+              />
+            }
+          />
         )}
 
         {actionItems.length > 0 && <ChatSessionActions actions={actionItems} />}
@@ -209,14 +212,16 @@ export const ChatMessageFooter = ({
         </div>
       ) : (
         msg.model && (
-          <SimpleTooltip
-            content={msg.model}
-            triggerRender={<span className="ml-auto min-w-0 shrink" />}>
-            <span className="inline-flex h-6 min-w-0 max-w-[clamp(5rem,24vw,14rem)] items-center gap-1 rounded-control px-1 text-[10px] text-muted-foreground/70 hover:bg-muted/35 hover:text-foreground">
-              <Bot className="icon-xs shrink-0" />
-              <span className="truncate">{msg.model}</span>
-            </span>
-          </SimpleTooltip>
+          <TooltipActionButton
+            trigger={<span className="ml-auto min-w-0 shrink" />}
+            tooltip={msg.model}
+            icon={
+              <span className="inline-flex h-6 min-w-0 max-w-[clamp(5rem,24vw,14rem)] items-center gap-1 rounded-control px-1 text-[10px] text-muted-foreground/70 hover:bg-muted/35 hover:text-foreground">
+                <Bot className="icon-xs shrink-0" />
+                <span className="truncate">{msg.model}</span>
+              </span>
+            }
+          />
         )
       )}
     </div>
