+++ src/features/chat/components/reasoning-trace.tsx
@@ -9,13 +9,13 @@ import {
 } from "lucide-react"
 import { useState } from "react"
 import { useTranslation } from "react-i18next"
+import { TooltipActionButton } from "@/components/actions"
 import { MarkdownRenderer } from "@/components/markdown-renderer"
 import {
   Popover,
   PopoverContent,
   PopoverTrigger
 } from "@/components/ui/popover"
-import { SimpleTooltip } from "@/components/ui/simple-tooltip"
 import {
   Tooltip,
   TooltipContent,
@@ -162,25 +162,29 @@ export const ReasoningTrace = ({
           const Icon = step.icon ?? Circle
           const label = getDisplayLabel(step.label, step.status)
           return (
-            <SimpleTooltip
+            <TooltipActionButton
               key={step.key}
-              content={label}
-              triggerRender={
+              trigger={
                 <span
                   className={cn(
                     "inline-flex size-7 items-center justify-center rounded-control transition-colors hover:bg-muted/45",
                     statusClass(step.status)
                   )}
                 />
-              }>
-              <Icon
-                className={cn(
-                  "icon-sm",
-                  step.status === "running" && "animate-pulse"
-                )}
-              />
-              <span className="sr-only">{label}</span>
-            </SimpleTooltip>
+              }
+              tooltip={label}
+              icon={
+                <>
+                  <Icon
+                    className={cn(
+                      "icon-sm",
+                      step.status === "running" && "animate-pulse"
+                    )}
+                  />
+                  <span className="sr-only">{label}</span>
+                </>
+              }
+            />
           )
         })}
       </div>
