+++ src/features/model/components/provider-version.tsx
@@ -1,7 +1,7 @@
 import { useTranslation } from "react-i18next"
 
+import { TooltipActionButton } from "@/components/actions"
 import { Badge } from "@/components/ui/badge"
-import { SimpleTooltip } from "@/components/ui/simple-tooltip"
 import { useProviderModels } from "@/features/model/hooks/use-provider-models"
 
 export const ProviderVersion = () => {
@@ -12,13 +12,15 @@ export const ProviderVersion = () => {
   if (versionError || !version) return null
   return (
     <div>
-      <SimpleTooltip
-        content={<p>{t("model.version.tooltip", { version })}</p>}
-        triggerRender={<div />}>
-        <Badge variant="outline" className="cursor-default px-4 py-2">
-          {version}
-        </Badge>
-      </SimpleTooltip>
+      <TooltipActionButton
+        trigger={<div />}
+        tooltip={<p>{t("model.version.tooltip", { version })}</p>}
+        icon={
+          <Badge variant="outline" className="cursor-default px-4 py-2">
+            {version}
+          </Badge>
+        }
+      />
     </div>
   )
 }
