+++ src/options/components/settings-page.tsx
@@ -19,8 +19,7 @@ import {
 import { SocialHandles } from "@/components/social-handles"
 import { SocialLinkButton } from "@/components/social-link-button"
 import { ThemeToggle } from "@/components/theme-toggle"
-import { ChatDisplaySettings } from "@/features/chat/components/chat-display-settings"
-import { SpeechSettings } from "@/features/chat/components/speech-settings"
+import { ChatDisplaySettings, SpeechSettings } from "@/features/chat/components"
 import { ContextSettings } from "@/features/context/components/context-settings"
 import { ContentExtractionSettings } from "@/features/model/components/content-extraction-settings"
 import { EmbeddingSettings } from "@/features/model/components/embedding-settings"
